refactor: simplify metadata progress update forms and remove unused imports
- Removed unused imports and components related to metadata progress updates.
- Consolidated logic for handling metadata updates into a single form component.
- Improved readability and maintainability of the dashboard code.
